% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.missoNet.R
\name{plot.cv.missoNet}
\alias{plot.cv.missoNet}
\title{Plot the cross-validation surface produced by cv.missoNet}
\usage{
\method{plot}{cv.missoNet}(x, type = c("cv.heatmap", "cv.scatter"), ...)
}
\arguments{
\item{x}{Fitted cv.missoNet object.}

\item{type}{Type of plot, can be either "cv.heatmap" or "cv.scatter". The default is "cv.heatmap".}

\item{...}{Other graphical parameters.}
}
\description{
Plot the cross-validation surface produced by cv.missoNet
}
