% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Generate simulation data from parameters.}
\usage{
generateData(
  X = NULL,
  Beta = NULL,
  E = NULL,
  Theta = NULL,
  n,
  p,
  q,
  rho,
  missing.type = "MCAR",
  Sigma.X = NULL,
  Beta.elem.sparsity = 0.2,
  Beta.row.sparsity = 0.2,
  with.seed = NULL
)
}
\arguments{
\item{X}{(Optional) a user supplied numeric predictor matrix (n by p). If not \code{NULL} then \code{Sigma.X} will be ignored.}

\item{Beta}{(Optional) a user supplied numeric coefficient matrix (p by q). If not \code{NULL} then \code{Beta.elem.sparsity} and \code{Beta.row.sparsity} will be ignored.}

\item{E}{(Optional) a user supplied numeric error matrix (n by q). If not \code{NULL} then \code{Theta} will be ignored. The final response matrix is generated by \code{Y = X \%*\% Beta + E}.}

\item{Theta}{(Optional) a user supplied inverse covariance matrix (q by q) for error. Default is \code{Theta = NULL} and the program will generate a block matrix having four blocks corresponding to four types of network structures: independent, weak graph, strong graph and chain. Only needed when \code{E = NULL}.}

\item{n}{Sample size.}

\item{p}{The dimensionality of predictors.}

\item{q}{The dimensionality of responses.}

\item{rho}{A scalar or a numeric vector of length q: a user supplied missing probability for response variables.}

\item{missing.type}{Character string: can be \code{"MCAR"}, \code{"MAR"} or \code{"MNAR"}; default is \code{missing.type = "MCAR"}.}

\item{Sigma.X}{A user supplied PD covariance matrix for generating multivariate normal distributed X. If \code{NULL}, use AR(1) covariance with 0.7 autocorrelation by default. Only needed when \code{X = NULL}.}

\item{Beta.elem.sparsity}{A value between 0 and 1 controlling the approximate proportion of non-zero elements in Beta. Only needed when \code{Beta = NULL}.}

\item{Beta.row.sparsity}{A value between 0 and 1 controlling the approximate proportion of non-zero rows in Beta. Only needed when \code{Beta = NULL}.}

\item{with.seed}{Random seed.}
}
\value{
This function returns a \code{list} containing:
\itemize{
\item \code{X}: predictor matrix (n by p).
\item \code{Y}: response matrix without missing data (n by q).
\item \code{Z}: response matrix with missing data (n by q).
\item \code{Beta}: coefficient matrix (p by q).
\item \code{Theta}: inverse covariance matrix (q by q)
\item \code{rho}: a vector of length q: working missing probability.
\item \code{missing.type}: type of missing data.
}
}
\description{
Generate simulation data from parameters.
}
\examples{
generateData(n = 300, p = 50, q = 20, rho = 0.1, missing.type = "MCAR")
}
