% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missoNet-package.R, R/missoNet.R
\docType{package}
\name{missoNet}
\alias{missoNet}
\title{missoNet: a package for multivariate regression and conditional network estimation with missing outputs.}
\usage{
missoNet(
  X,
  Y,
  lambda.Beta,
  lambda.Theta,
  rho = NULL,
  Beta.maxit = 10000,
  Beta.thr = 1e-05,
  Theta.maxit = 10000,
  Theta.thr = 1e-05,
  eps = 1e-08,
  diag.penalty.factor = NULL,
  standardize = TRUE,
  standardize.response = TRUE,
  fit.relax = FALSE,
  verbose = 1
)
}
\arguments{
\item{X}{Numeric predictor matrix (n by p): columns correspond to predictor variables and rows correspond to samples. Missing values are not allowed. Do not include a column of ones.}

\item{Y}{Numeric response matrix (n by q): columns correspond to response variables and rows correspond to samples. Missing values should be coded as \code{NA} or \code{NaN}.}

\item{lambda.Beta}{Numeric: a user supplied positive value for regularizing \code{Beta}.}

\item{lambda.Theta}{Numeric: a user supplied positive value for regularizing \code{Theta}.}

\item{rho}{(Optional) A scalar or a numeric vector of length q: a user supplied missing probability for response variables. Default is \code{rho = NULL} and the program will compute the empirical missing rates for columns of \code{Y} and use them as the missing probability.}

\item{Beta.maxit}{The maximum number of internal iterations allowed for updating \code{Beta}. Default is \code{Beta.maxit = 1e4}.}

\item{Beta.thr}{The convergence threshold for updating \code{Beta}; default is \code{Beta.thr = 1e-5}. Iterations stop when absolute parameter change is less than \code{Beta.thr * sum(abs(Beta))}.}

\item{Theta.maxit}{The maximum number of internal iterations allowed for updating \code{Theta}. Default is \code{Theta.maxit = 1e4}.}

\item{Theta.thr}{The convergence threshold for updating \code{Theta}; default is \code{Theta.thr = 1e-5}. Iterations stop when average absolute parameter change is less than \code{Theta.thr * ave(abs(offdiag(Sigma)))}.}

\item{eps}{A numeric tolerance level for L1 projection; default is \code{eps = 1e-8}. If any of the eigenvalues is less than the given tolerance, the unbiased estimate of covariance is projected onto L1 ball to have \code{min(eigen(Sigma)$value) == eps}.}

\item{diag.penalty.factor}{Numeric: a separate penalty factor for the diagonal entries of \code{Theta}. This is a number that multiplies \code{lambda.Theta} to allow differential shrinkage. Default is \code{NULL} and the program computes it based on an initial estimate of \code{Theta}. Can be 0 for no shrinkage. Only needed when n <= p.}

\item{standardize}{Logical: should the columns of \code{X} be standardized so each has unit length and zero average; default is \code{TRUE}. The parameter estimates will be returned on the original scale. If \code{X} has been standardized prior to fitting the model, you might not wish to standardize.}

\item{standardize.response}{Logical: should the columns of \code{Y} be standardized so each has unit length and zero average; default is \code{TRUE}. The parameter estimates will be returned on the original scale. If \code{Y} has been standardized prior to fitting the model, you might not wish to standardize.}

\item{fit.relax}{Logical: default is \code{FALSE}. If \code{TRUE}, the program will re-estimate a relaxed graph (\code{Theta}) without penalization, which could be useful for network analysis. WARNING: there may be convergence issues if the residual covariance matrix is not of full rank.}

\item{verbose}{Value of 0, 1 or 2. 0 -- silent; 1 -- limited tracing; 2 -- detailed tracing.}
}
\value{
This function returns a \code{list} of estimates and working parameters:
\itemize{
\item \code{est}: estimates using the designated regularization parameters.
\item \code{relax.graph}: an estimate of network without penalization if \code{fit.relax = TRUE}.
\item \code{rho}: a vector of length q: working missing probability.
\item \code{penalize.diagonal}: logical: whether the diagonal of \code{Theta} was penalized.
\item \code{diag.penalty.factor}: penalty factor for the diagonal when \code{penalize.diagonal} returned as \code{TRUE}.
}
}
\description{
The missoNet package provides three categories of important functions:
cv.missoNet, plot/predict and generateData.
}
\section{missoNet functions}{

cv.missoNet: fit missoNet with the optimal tuning parameters selected by a k-fold cross-validation.
plot: S3 method for plotting the cross-validation errors from a "cv.missoNet" object.
predict: S3 method for making predictions from a "cv.missoNet" object.
generateData: generate simulation data by specifying a certain type of missing mechanism.
}

