% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.missoNet.R
\name{plot.cv.missoNet}
\alias{plot.cv.missoNet}
\title{Plot the cross-validation errors produced by \code{cv.missoNet}}
\usage{
\method{plot}{cv.missoNet}(x, type = c("cv.heatmap", "cv.scatter"), detailed.axes = TRUE, ...)
}
\arguments{
\item{x}{Fitted `\code{cv.missoNet}` object.}

\item{type}{Type of plot, can be either `\code{"cv.heatmap"}` (default) or `\code{"cv.scatter"}`.}

\item{detailed.axes}{Logical: should the detailed axes be plotted? The default is `TRUE`.}

\item{...}{Other graphical arguments used by `\code{ComplexHeatmap::Heatmap}`.}
}
\description{
S3 method for plotting the mean cross-validation error surface from a fitted `\code{cv.missoNet}` object.
}
