% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cv.missoNet.R
\name{predict.cv.missoNet}
\alias{predict.cv.missoNet}
\title{Make predictions from a \code{cv.missoNet} object}
\usage{
\method{predict}{cv.missoNet}(object, newx = NULL, s = "lambda.min", ...)
}
\arguments{
\item{object}{Fitted `\code{cv.missoNet}` object.}

\item{newx}{A predictor matrix of new values at which predictions are to be made. `\code{newx}` should have the same scale as the training data. Missing values are not allowed and do not include a column of ones.}

\item{s}{Character, the penalty parameter `lambda` at which the regression coefficients for predictions are extracted. It supports three special values, named `\code{"lambda.min"}` (default), `\code{"lambda.Beta.1se"}` and `\code{"lambda.Theta.1se"}`.}

\item{...}{Not used. Other arguments for predicting.}
}
\value{
Predicted response matrix `\code{newy}`.
}
\description{
S3 method for making predictions from a fitted `\code{cv.missoNet}` object.
}
