% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cv.missoNet.R
\name{predict.cv.missoNet}
\alias{predict.cv.missoNet}
\title{Make predictions from a cv.missoNet object.}
\usage{
\method{predict}{cv.missoNet}(object, newx = NULL, s = "lambda.min", ...)
}
\arguments{
\item{object}{Fitted cv.missoNet object.}

\item{newx}{Matrix of new values for X at which predictions are to be made.}

\item{s}{The penalty parameter lambda at which predictions are required. Default is \code{s = "lambda.min"} stored in the CV object. Alternatively \code{s = "lambda.1se"} can be used.}

\item{...}{Not used. Other arguments for predicting.}
}
\value{
Predicted response matrix code{newY}.
}
\description{
Make predictions from a cv.missoNet object.
}
